pyproject.toml
[tool.poetry]
name = "microGaitMs"
version = "0.1.0"
description = "Package to deal with micro Gait expressions in Multiple Sclerosis"
authors = ["Angela Fernandez <angela.fernandez@alumnos.upm.es>"]
packages = [{ include = "InfluxDBms" },
            { include = "msGeom"}]
license = "GPL 3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
influxdb_client = "^1.48.0"
openpyxl = "^3.1.5"
matplotlib = "<3.9.1"
pandas = "^2.2.3"
numpy = "^2.1"
PyYAML = "^6.0.2"
pyproj = ">=3.6.0,<4"
filterpy = ">=1.4.0,<2"
plotly = ">=6.0.1,<7"
setuptools = "^75.8.0"
urllib3 = "^2.4.0"
folium = "^0.18"
gait-tracking = { path = "../external_repos/Gait-Tracking-With-x-IMU-Python" }
scikit-kinematics = { path = "../external_repos/scikit-kinematics" }
tables = "^3.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
